final version using Spring Data JDBC

+++ src/main/java/com/minhajcse/service/UserService.java
@@ -1,6 +1,8 @@
 package com.minhajcse.service;
 
+import com.minhajcse.model.Author;
 import com.minhajcse.model.User;
+import com.minhajcse.repository.AuthorRepository;
 import com.minhajcse.repository.UserRepository;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
@@ -10,8 +12,10 @@ import java.util.List;
 @Service
 public class UserService {
     private final UserRepository userRepository;
-    UserService(UserRepository userRepository) {
+    private final AuthorRepository authorRepository;
+    UserService(UserRepository userRepository, AuthorRepository authorRepository) {
         this.userRepository = userRepository;
+        this.authorRepository = authorRepository;
     }
 
     public User getUserById(Long id) {
@@ -26,8 +30,13 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+
+    public void deleteUser(Long id) throws Exception {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else {
+            throw new Exception("User with id " + id + " not found");
+        }
     }
 
     public User updateUser(User user) {
